+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Added support to set the size of an Order - [ripe-pulse/#389](https://github.com/ripe-tech/ripe-pulse/issues/389)
+* Added support to convert sizes from native to human in bulk to specific scale - [ripe-pulse/#389](https://github.com/ripe-tech/ripe-pulse/issues/389)
 
 ### Changed